Add Fabric Support for uploaded mods.
There might not be much in terms of service, but I do know that Fabric offers a nice little .json file to probe. Perhaps if there is a good amount of support, we can find where that version number is hidden. (For now, however, it's hard coded as f1.15.2- What I use to distinguish between Forge 1.15.2, and Fabric 1.15.2.).

Some mods do use the "Minecraft" depends feature, but not all, which makes it quite annoying.